Compiled .EXE file
This file is for my purpose only. The network drive references will not work on anyone else's environment, but it will not bomb out if you give it a try. The program will just seem to "hang" up for a bit when trying to run the scripts. No error will show.